requirements/custom.txt
PIL==1.1.7
South==0.7.6
xlrd==0.9.2
johnny-cache==1.4
pylibmc==1.2.3
gunicorn==0.17.4
raven==3.3.7
MySQL-python==1.2.4
facebook-sdk==0.4.0
django-social-auth==0.7.27
premailer==2.0.2
